+++ dwds/lib/src/dwds_vm_client.dart
@@ -5,10 +5,13 @@
 import 'dart:async';
 import 'dart:convert';
 
+import 'package:logging/logging.dart';
 import 'package:vm_service/vm_service.dart';
 
 import 'services/chrome_proxy_service.dart' show ChromeProxyService;
 import 'services/debug_service.dart';
+
+final _logger = Logger('DwdsVmClient');
 
 // A client of the vm service that registers some custom extensions like
 // hotRestart.
@@ -31,10 +34,16 @@ class DwdsVmClient {
     var responseController = StreamController<Map<String, Object>>();
     VmServerConnection(requestController.stream, responseController.sink,
         debugService.serviceExtensionRegistry, debugService.chromeProxyService);
-    var client = VmService(
-        responseController.stream.map(jsonEncode),
-        (request) => requestController.sink
-            .add(jsonDecode(request) as Map<String, dynamic>));
+    var client =
+        VmService(responseController.stream.map(jsonEncode), (request) {
+      if (requestController.isClosed) {
+        _logger.warning(
+            'Attempted to send a request but the connection is closed:\n\n'
+            '$request');
+        return;
+      }
+      requestController.sink.add(jsonDecode(request) as Map<String, dynamic>);
+    });
     var chromeProxyService =
         debugService.chromeProxyService as ChromeProxyService;